added editing and deleting functions + functional editing modals

+++ src/blockly/serialization.js
@@ -34,7 +34,21 @@ export function loadBlockly(name, workspace) {
 export function saveBlockly(name, workspace) {
     const data = Blockly.serialization.workspaces.save(workspace)
     let projects = getProjects()
-    console.log(1)
     projects[name] = { time: dayjs(), workspace: data }
+    localStorage.setItem("roboxProjects", JSON.stringify(projects))
+}
+export function editProject(oldName, newName) {
+    let projects = getProjects()
+    if (!projects[oldName]) throw new Error("Project does not exists")
+    if (projects[newName]) throw new Error("Project already exists")
+    projects[newName] = structuredClone(projects[oldName])
+    delete projects[oldName]
+    console.log(1)
+    localStorage.setItem("roboxProjects", JSON.stringify(projects))
+}
+export function deleteProject(name) {
+    let projects = getProjects()
+    if (!projects[name]) throw new Error("Project does not exists")
+    delete projects[name]
     localStorage.setItem("roboxProjects", JSON.stringify(projects))
 }
